Adding tests, passing all but...
there is an issue with cluster validation on the training of fake seed 201.
also on the seeded tests there are less clusters overall

this results in way worse results
seed 200 had (2, 6, 10) clusters and now is (2, 2, 2) resulting in p,n,u from (80,18,01) to  (77,22,00)
(dont even mention forest cover)

Also, tests of dask are missng and multiprocessing impl. is yet to start.